Addresses comments in PR
